Move everything to /expressions/
