+++ dali/pipeline/operators/expressions/arithmetic_meta.h
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef DALI_PIPELINE_OPERATORS_ARITHMETIC_ARITHMETIC_META_H_
-#define DALI_PIPELINE_OPERATORS_ARITHMETIC_ARITHMETIC_META_H_
+#ifndef DALI_PIPELINE_OPERATORS_EXPRESSIONS_ARITHMETIC_META_H_
+#define DALI_PIPELINE_OPERATORS_EXPRESSIONS_ARITHMETIC_META_H_
 
 #include <cstdint>
 #include <map>
@@ -342,4 +342,4 @@ inline bool IsScalarLike(const kernels::TensorListShape<> &shape) {
 
 }  // namespace dali
 
-#endif  // DALI_PIPELINE_OPERATORS_ARITHMETIC_ARITHMETIC_META_H_
+#endif  // DALI_PIPELINE_OPERATORS_EXPRESSIONS_ARITHMETIC_META_H_
