+++ dali/pipeline/operators/expressions/expression_impl_cpu.h
@@ -12,16 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef DALI_PIPELINE_OPERATORS_ARITHMETIC_EXPRESSION_IMPL_CPU_H_
-#define DALI_PIPELINE_OPERATORS_ARITHMETIC_EXPRESSION_IMPL_CPU_H_
+#ifndef DALI_PIPELINE_OPERATORS_EXPRESSIONS_EXPRESSION_IMPL_CPU_H_
+#define DALI_PIPELINE_OPERATORS_EXPRESSIONS_EXPRESSION_IMPL_CPU_H_
 
 #include <vector>
 
 #include "dali/core/any.h"
 #include "dali/pipeline/data/types.h"
-#include "dali/pipeline/operators/arithmetic/arithmetic_meta.h"
-#include "dali/pipeline/operators/arithmetic/expression_impl_factory.h"
-#include "dali/pipeline/operators/arithmetic/expression_tree.h"
+#include "dali/pipeline/operators/expressions/arithmetic_meta.h"
+#include "dali/pipeline/operators/expressions/expression_impl_factory.h"
+#include "dali/pipeline/operators/expressions/expression_tree.h"
 #include "dali/pipeline/operators/op_spec.h"
 #include "dali/pipeline/util/backend2workspace_map.h"
 #include "dali/pipeline/workspace/workspace.h"
@@ -74,4 +74,4 @@ class ExprImplBinCPU : public ExprImplBase, ExprImplParam<CPUBackend> {
 
 }  // namespace dali
 
-#endif  // DALI_PIPELINE_OPERATORS_ARITHMETIC_EXPRESSION_IMPL_CPU_H_
+#endif  // DALI_PIPELINE_OPERATORS_EXPRESSIONS_EXPRESSION_IMPL_CPU_H_
