+++ dali/pipeline/operators/expressions/expression_impl_gpu.cuh
@@ -12,16 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef DALI_PIPELINE_OPERATORS_ARITHMETIC_EXPRESSION_IMPL_GPU_CUH_
-#define DALI_PIPELINE_OPERATORS_ARITHMETIC_EXPRESSION_IMPL_GPU_CUH_
+#ifndef DALI_PIPELINE_OPERATORS_EXPRESSIONS_EXPRESSION_IMPL_GPU_CUH_
+#define DALI_PIPELINE_OPERATORS_EXPRESSIONS_EXPRESSION_IMPL_GPU_CUH_
 
 #include <vector>
 
 #include "dali/core/any.h"
 #include "dali/pipeline/data/types.h"
-#include "dali/pipeline/operators/arithmetic/arithmetic_meta.h"
-#include "dali/pipeline/operators/arithmetic/expression_impl_factory.h"
-#include "dali/pipeline/operators/arithmetic/expression_tree.h"
+#include "dali/pipeline/operators/expressions/arithmetic_meta.h"
+#include "dali/pipeline/operators/expressions/expression_impl_factory.h"
+#include "dali/pipeline/operators/expressions/expression_tree.h"
 #include "dali/pipeline/operators/op_spec.h"
 #include "dali/pipeline/util/backend2workspace_map.h"
 #include "dali/pipeline/workspace/workspace.h"
@@ -115,4 +115,4 @@ class ExprImplBinGPU : public ExprImplBase, ExprImplParam<GPUBackend> {
 
 }  // namespace dali
 
-#endif  // DALI_PIPELINE_OPERATORS_ARITHMETIC_EXPRESSION_IMPL_GPU_CUH_
+#endif  // DALI_PIPELINE_OPERATORS_EXPRESSIONS_EXPRESSION_IMPL_GPU_CUH_
